% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{add_residuals}
\alias{add_residuals}
\title{Returns a melted data table of species and their scaled cmps (based on model coefficient), along with residual contributions}
\usage{
add_residuals(species_cmps, model_dat, resid_dat)
}
\arguments{
\item{species_cmps}{data table of species and their CMPs}

\item{model_dat}{Table of model results (from fit_cmp_mods)}

\item{resid_dat}{Table of model residuals (from fit_cmp_mods)}
}
\value{
A melted data table of species and their CMPs, including "Residual" as an additional species
}
\description{
Returns a melted data table of species and their scaled cmps (based on model coefficient), along with residual contributions
}
\examples{
add_residuals(species_cmps, mod_results[[1]], mod_results[[2]])
}
