% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{add_rev_rxns}
\alias{add_rev_rxns}
\title{Add reverse of reversible reactions to edge list network}
\usage{
add_rev_rxns(network, refine = F)
}
\arguments{
\item{network}{community metabolic network model}
}
\value{
network with reverse of reversible reactions added, with separate reaction IDs (_1 added to ID)
}
\description{
Add reverse of reversible reactions to edge list network
}
\examples{
add_rev_rxns(rxn_table)
}
