% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{add_rxns_to_network}
\alias{add_rxns_to_network}
\title{Add reactions to a network}
\usage{
add_rxns_to_network(network, netAddTable)
}
\arguments{
\item{network}{Data.table of taxa, genes and reactions}

\item{netAdd}{Data.table of taxa, genes and reactions to add, or generic genes and reactions to be applied to all taxa}
}
\value{
Expanded network table
}
\description{
Add reactions to a network
}
\examples{
add_rxns_to_network(network, netAddTable)
}
