% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{add_to_network}
\alias{add_to_network}
\title{Add reactions to a network. Will set stoichiometry and copy number to 1 if missing. Format can be either "KO, Rxn, Prod" for reaction IDs with all transformations and correct stoichiometry, or just "KO" but with reaction IDs that are defined in the KEGG network.}
\usage{
add_to_network(network, addTable, target_format = NULL,
  source_format = NULL, kegg_path = "data/KEGGfiles/",
  data_path = "data/")
}
\arguments{
\item{network}{Data.table of taxa, genes and reactions}

\item{addTable}{Data.table of taxa, genes and/or reactions to add, or generic genes and reactions to be applied to all taxa}

\item{target_format}{Format of taxa, genes, and/or reactions to add - must be "KEGG" or "Cobra". If NULL, will try to guess}

\item{source_format}{Format of taxa, genes, and/or reactions to add - must be "KEGG" or "Cobra". If NULL, will try to guess}
}
\value{
Expanded network table
}
\description{
Add reactions to a network. Will set stoichiometry and copy number to 1 if missing. Format can be either "KO, Rxn, Prod" for reaction IDs with all transformations and correct stoichiometry, or just "KO" but with reaction IDs that are defined in the KEGG network.
}
\examples{
add_to_network(network, netAddTable)
}
