% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{agora_kegg_mets}
\alias{agora_kegg_mets}
\title{Finds KEGG IDs for a list of AGORA metabolites}
\usage{
agora_kegg_mets(agora_ids)
}
\arguments{
\item{agora_ids}{List of AGORA metabolite IDs}
}
\value{
vector of KEGG IDs corresponding to agora_ids
}
\description{
Finds KEGG IDs for a list of AGORA metabolites
}
\examples{
agora_kegg_mets(agora_ids)
}
