% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{basic_correlation_matrix}
\alias{basic_correlation_matrix}
\title{Basic species-metaoblite correlation for comparison}
\usage{
basic_correlation_matrix(species, mets, method = "pearson")
}
\arguments{
\item{species}{long-form table of species abundances}

\item{mets}{long-form table of metabolite concentrations}

\item{method}{Correlation method, must be one recognized by cor.test}
}
\value{
Table of correlation coefficients and p-values for every species-metabolite pair with sufficient complete sample data
}
\description{
Basic species-metaoblite correlation for comparison
}
