% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{blast_seqs}
\alias{blast_seqs}
\title{BLASTS a set of sequences against a database}
\usage{
blast_seqs(seqs, blast_path = "data/blastDB/")
}
\arguments{
\item{seqs}{Biostrings object of sequences}

\item{blast_path}{Path to BLAST directory}
}
\value{
BLAST results
}
\description{
BLASTS a set of sequences against a database
}
\examples{
blast_seqs(seqs, blast_path = "data/blastDB")
}
