% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestOR.R
\name{Boxplot}
\alias{Boxplot}
\title{Boxplots from MIMOSA fit}
\usage{
Boxplot(obj, truth)
}
\arguments{
\item{obj}{\code{MIMOSA2()} output}

\item{truth}{\code{character vector} of true responder / non-responder calls. Responder should be "R"}
}
\value{
ggplot grob
}
\description{
Boxplots from MIMOSA fit
}
\examples{
s = simulate_MIMOSA2()
R = MIMOSA2(Ntot=s$Ntot, ns1 = s$ns1, nu1 = s$nu1, nu0 = s$nu0, ns0 = s$ns0,maxit=10)
Boxplot(obj = R,truth = s$truth)
}
\seealso{
\link{MIMOSA2}
}
