% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{build_metabolic_model}
\alias{build_metabolic_model}
\title{Build species-specific metabolic model for MIMOSA2 analysis}
\usage{
build_metabolic_model(species, config_table,
  gg_path = "data/picrustGenomeData/indivModels/", netAdd = NULL)
}
\arguments{
\item{species}{Table of species/taxon abundances}

\item{config_table}{Data.table of input files and settings for MIMOSA}

\item{gg_path}{Path to Picrust ref files}

\item{netAdd}{Table of netowrk information to add to the model}
}
\value{
Data.table of network model of genes and reactions for each species/taxon
}
\description{
Build species-specific metabolic model for MIMOSA2 analysis
}
\examples{
build_metabolic_model(config_table)
}
