% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{calculate_var_shares_linear}
\alias{calculate_var_shares_linear}
\title{Calculates contributions to variance from a contribution table}
\usage{
calculate_var_shares_linear(species_contribution_table,
  valueVar = "newValue", model_cov = F)
}
\arguments{
\item{species_contribution_table}{Table of species contributions (with residuals)}

\item{valueVar}{Column name to use for values}

\item{model_cov}{Just calculate covariances with metabolite/response data}

\item{metVar}{If model_cov = T, variable name for metabolite concentrations}
}
\value{
Data table of variance shares by each species in the original table for each metabolite
}
\description{
Calculates contributions to variance from a contribution table
}
\examples{
calculate_var_shares_linear(species_cmps)
}
