% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{check_ref_data}
\alias{check_ref_data}
\title{Check that MIMOSA2 can access reference databases in the correct format}
\usage{
check_ref_data(seq_db, target_db)
}
\arguments{
\item{seq_db}{}

\item{target_db}{}
}
\value{
T if ref data is accessible as expected, F if not
}
\description{
Check that MIMOSA2 can access reference databases in the correct format
}
\examples{
check_ref_data("Sequence variants (ASVs)", "AGORA genomes and models")
}
