% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{cmp_met_plot}
\alias{cmp_met_plot}
\title{Plot CMP values vs metabolite concentrations for a single metabolite}
\usage{
cmp_met_plot(
  cmp_table,
  met_table,
  mod_results = NULL,
  sample_col = "Sample",
  met_title = NULL
)
}
\arguments{
\item{cmp_table}{Dataset of CMP scores (long format, can be species-specific or added already) for a single metabolite}

\item{met_table}{Dataset of metabolite concentrations (long format) for a single metabolite}

\item{mod_results}{Optional additional info on model comparison results}

\item{sample_col}{Column name for sample column (shared between the tables)}
}
\value{
plot of contributions
}
\description{
Plot CMP values vs metabolite concentrations for a single metabolite
}
\examples{
cmp_met_plot(cmp_scores, mets_melt, mod_results = mods_out[[1]])
}
