% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{correct}
\alias{correct}
\title{#' Title
#'
#' @param nrow 
#' @param ncol 
#'
#' @return
#' @export
#'
#' @examples
make_perm_mat2 = function(nrow, ncol){
  make_perm_mat(nrow, ncol)
}
Multiple hypothesis correction}
\usage{
correct(pvals, method = "fdr")
}
\arguments{
\item{pvals}{Vector of p-values}

\item{method}{Must be either "fdr" or "bonferroni"}
}
\value{
Vector of corrected values
}
\description{
#' Title
#'
#' @param nrow 
#' @param ncol 
#'
#' @return
#' @export
#'
#' @examples
make_perm_mat2 = function(nrow, ncol){
  make_perm_mat(nrow, ncol)
}
Multiple hypothesis correction
}
\examples{

}
