% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{download_reference_data}
\alias{download_reference_data}
\title{Uses the provided OTU table to generate a contribution table based on the PICRUSt 16S normalization and genomic content tables}
\usage{
download_reference_data(
  seq_db = get_text("database_choices")[1],
  target_db = get_text("source_choices")[2],
  save_to = paste0(getwd(), "/"),
  source_url = "http://elbo-spice.cs.tau.ac.il/shiny/MIMOSA2shiny/refData/"
)
}
\arguments{
\item{seq_db}{Microbiome data source type: must be one of "Sequence variants (ASVs)", "Greengenes 13_5 or 13_8 OTUs", or "SILVA 132 OTUs"}

\item{target_db}{Metabolic network source: must be one of "AGORA genomes and models" or "RefSeq/EMBL_GEMs genomes and models" (for KEGG models, see documentation)}

\item{save_to}{File path to save the downloaded data. Default is a folder named "data/" in the current working directory}
}
\value{
None
}
\description{
Uses the provided OTU table to generate a contribution table based on the PICRUSt 16S normalization and genomic content tables
}
\examples{
download_reference_data("Sequence variants (ASVs)", "AGORA genomes and models")
}
