% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{emm_to_edge_list}
\alias{emm_to_edge_list}
\title{Convert stoichiometric matrix to edge list of reactions}
\usage{
emm_to_edge_list(emm)
}
\arguments{
\item{emm}{Stoichiometric matrix (rows are metabolites, columns are reactions)}
}
\value{
A data.table where each row is a pair of compounds being exchanged in a reaction
}
\description{
Convert stoichiometric matrix to edge list of reactions
}
\examples{
emm_to_edge_list(s_mat)
}
