% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{filter_species_abunds}
\alias{filter_species_abunds}
\title{Filter species found in few samples and/or low abundance}
\usage{
filter_species_abunds(
  species_dat,
  filter_type = "mean",
  minMeanAbund = 0,
  minSampFrac = 0
)
}
\arguments{
\item{species_dat}{OTU table}

\item{filter_type}{Either "mean" or "fracNonzero"}

\item{minMeanAbund}{minimum mean relative abundance for inclusion of a taxon}

\item{minSampFrac}{minimum fraction of samples with nonzero abundance for inclusion of a taxon}
}
\value{
Filtered species table
}
\description{
Filter species found in few samples and/or low abundance
}
\examples{
filter_species_abunds(species_dat)
}
