% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{fit_single_scaling_mod}
\alias{fit_single_scaling_mod}
\title{Fits single model scaling total CMPs to concentrations of a metabolite}
\usage{
fit_single_scaling_mod(met1, cmp_rxn_dat, mets_melt)
}
\arguments{
\item{met1}{metabolite ID to fit}

\item{cmp_rxn_dat}{Table of species-reaction contribution abundances}

\item{mets_melt}{Table of metabolite concentrations}
}
\value{
lm model object
}
\description{
Fits single model scaling total CMPs to concentrations of a metabolite
}
\examples{
fit_single_scaling_mod(met1, species_cmps, met_data)
}
