% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{g_legend}
\alias{g_legend}
\title{Get legend from a ggplot2 object}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{}
}
\value{
ggplot2 legend object
}
\description{
Get legend from a ggplot2 object
}
\examples{
g_legend(my_plot)
}
