% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{generate_preprocessed_networks}
\alias{generate_preprocessed_networks}
\title{Generate preprocessed reference databases for MIMOSA2}
\usage{
generate_preprocessed_networks(
  database,
  model_table_file = NULL,
  picrust_ko_path = "data/picrustGenomeData/",
  kegg_paths = c("data/KEGG/reaction_mapformula.lst", "data/KEGG/reaction_ko.list",
    "data/KEGG/reaction"),
  dat_path = paste0("data/", database, "/"),
  out_path = paste0("data/", database, "/RxnNetworks/")
)
}
\arguments{
\item{database}{One of "KEGG", "picrustGG", "AGORA", or "embl_gems"}

\item{model_table_file}{Optionally, a table listing models from a version of AGORA or embl_gems other than what is included in package data}

\item{picrust_ko_path}{Path to unzipped precalculated PICRUSt 1 files}

\item{kegg_paths}{Vector of file paths to the 3 KEGG files required for MIMOSA network template construction (1) reaction_mapformula.lst, 2) reaction_ko.list, and 3) reaction)}

\item{dat_path}{For the AGORA or embl_gems options, path to directory of .mat model files}

\item{out_path}{Path to write processed reaction network outputs for each taxon}
}
\value{
No return, writes processed reaction network to file
}
\description{
Generate preprocessed reference databases for MIMOSA2
}
\examples{
generate_preprocessed_networks("AGORA", dat_path = "data/AGORA/", )
}
