% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{get_agora_from_otus}
\alias{get_agora_from_otus}
\title{Returns AGORA IDs that have been previously mapped to a set of closed-reference OTU IDs}
\usage{
get_agora_from_otus(otus, database = database_choices[2],
  data_path = "data/rep_seqs/agora_mapping/")
}
\arguments{
\item{otus}{List of OTU IDs}

\item{database}{Database/method used for generating populations in the table - currently Greengenes or SILVA}

\item{data_path}{File path to OTU-AGORA mapping}
}
\value{
Biostrings object containing representative sequences for each OTU in order
}
\description{
Returns AGORA IDs that have been previously mapped to a set of closed-reference OTU IDs
}
\examples{
get_rep_seqs_from_otus(otu_list, "Greengenes 13_5 or 13_8", "data/rep_seqs/")
}
