% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{get_analysis_summary}
\alias{get_analysis_summary}
\title{Calculate general summary statistics on the results of a MIMOSA analysis}
\usage{
get_analysis_summary(
  input_species,
  species,
  mets,
  network,
  indiv_cmps,
  cmp_mods,
  var_shares,
  config_table,
  pval_threshold = 0.1
)
}
\arguments{
\item{species}{}

\item{mets}{}

\item{indiv_cmps}{}

\item{cmp_mods}{}

\item{var_shares}{}

\item{config_table}{}
}
\value{
Table of summary statistics (1st column is description, 2nd column is value)
}
\description{
Calculate general summary statistics on the results of a MIMOSA analysis
}
\examples{
get_analysis_summary(input_species, species, mets, network, cmps, mod_results, contributions, config_table)
}
