% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{get_cmp_scores_kos}
\alias{get_cmp_scores_kos}
\title{Updated version of getting all sample-level CMP scores from a KO abundance table}
\usage{
get_cmp_scores_kos(
  ko_table,
  network,
  normalize = T,
  relAbund = T,
  remove_rev = T,
  leave_rxns = F
)
}
\arguments{
\item{ko_table}{KO abundance table (wide format)}

\item{network}{Species-specific network table, product of build_network functions}

\item{normalize}{Whether to normalize rows when making the network EMM}

\item{remove_rev}{Whether to remove rev rxns before calculating/normalizing}

\item{leave_rxns}{Whether to leave individual KO contributions or aggregate to metabolite level}
}
\value{
data.table of cmp scores for each taxon and compound
}
\description{
Updated version of getting all sample-level CMP scores from a KO abundance table
}
\examples{
get_cmp_scores_kos(ko_data, network)
}
