% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{get_cmp_summary}
\alias{get_cmp_summary}
\title{Get summary of CMP score basis across all samples}
\usage{
get_cmp_summary(species_table, network, normalize = T, relAbund = T,
  manual_agora = F, humann2 = F, kos_only = F, remove_rev = T,
  met_subset = NULL, contrib_sizes = NULL, return_var_shares = T)
}
\arguments{
\item{species_table}{OTU abundance table (wide format)}

\item{network}{Species-specific network table, product of build_network functions}

\item{normalize}{Whether to normalize rows when making the network EMM}

\item{relAbund}{Whether to use relative abundance normalization}

\item{manual_agora}{Whether the metabolite are already specified using AGORA/BiGG (and therefore should not be mapped back to KEGG)}

\item{humann2}{Whether the species data is long-form humann2 gene-species abundances}

\item{kos_only}{Whether to call non-species-specific version of this function instead}

\item{remove_rev}{Whether to remove reversible reactions before calculating anything}

\item{met_subset}{Subset of metabolites to keep}

\item{contrib_sizes}{Contribution values to order by, if applicable}

\item{return_var_shares}{Whether to also return contributions table with summary information}
}
\value{
List of tables of stats on relevant nonzero species and reactions in the network for each compound
}
\description{
Get summary of CMP score basis across all samples
}
\examples{
get_cmp_summary(species, network)
}
