% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{get_compound_format}
\alias{get_compound_format}
\title{Determines whether a list of compounds is in KEGG or AGORA format or neither}
\usage{
get_compound_format(comp_list)
}
\arguments{
\item{comp_list}{List of compound IDs}
}
\value{
Either "KEGG" or "AGORA"
}
\description{
Determines whether a list of compounds is in KEGG or AGORA format or neither
}
\examples{
get_compound_format(met_ids)
}
