% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{get_genomic_content_from_picrust_table}
\alias{get_genomic_content_from_picrust_table}
\title{Returns the column from the picrust tables that corresponds to the genomic content of the indicated OTU}
\usage{
get_genomic_content_from_picrust_table(otu, picrust_ko_table_directory,
  picrust_ko_table_suffix)
}
\arguments{
\item{otu}{OTU ID}

\item{picrust_ko_table_directory}{Directory of PICRUSt genome OTU predictions}

\item{picrust_ko_table_suffix}{File naming of PICRUSt genome OTU predictions}
}
\value{
Genomic content for a single OTU
}
\description{
Returns the column from the picrust tables that corresponds to the genomic content of the indicated OTU
}
\examples{
get_genomic_content_from_picrust_table(otu, picrust_dir, picrust_suffix)
}
