% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{get_kegg_met_ids}
\alias{get_kegg_met_ids}
\title{Finds KEGG IDs for a list of AGORA metabolites}
\usage{
get_kegg_met_ids(agora_ids, kegg_table)
}
\arguments{
\item{agora_ids}{List of AGORA metabolite IDs}

\item{kegg_table}{Dictionary of AGORA-KEGG IDs}
}
\value{
vector of KEGG IDs corresponding to agora_ids
}
\description{
Finds KEGG IDs for a list of AGORA metabolites
}
\examples{
get_kegg_met_ids(agora_ids, kegg_table)
}
