% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{get_rep_seqs_from_otus}
\alias{get_rep_seqs_from_otus}
\title{Returns representative 16S sequences for a set of closed-reference OTU IDs}
\usage{
get_rep_seqs_from_otus(
  otus,
  database = database_choices[2],
  rep_seq_path = "data/rep_seqs/"
)
}
\arguments{
\item{otus}{List of OTU IDs}

\item{database}{Database/method used for generating populations in the table - currently Greengenes or SILVA}

\item{rep_seq_path}{File path to OTU representative sequences}
}
\value{
Biostrings object containing representative sequences for each OTU in order
}
\description{
Returns representative 16S sequences for a set of closed-reference OTU IDs
}
\examples{
get_rep_seqs_from_otus(otu_list, "Greengenes 13_5 or 13_8", "data/rep_seqs/")
}
