% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{get_S_mats}
\alias{get_S_mats}
\title{Convert stoichiometric matrix to edge list of reactions}
\usage{
get_S_mats(
  all_mods,
  species_names,
  edge_list = F,
  get_bounds = T,
  remove_sinks = T
)
}
\arguments{
\item{all_mods}{List of Cobra-formatted metabolic models}

\item{species_names}{List of species names for each model}

\item{edge_list}{Whether to return the metabolic network as a stoichiometric matrix or edge list}

\item{get_bounds}{Whether to extract information on upper and lower bound constraints (only with edge list)}

\item{remove_sinks}{Remove reactions that have NA as a substrate or product (usually uninformative transporters etc)}
}
\value{
Either a stoichiometric matrix or edge list of reactions
}
\description{
Convert stoichiometric matrix to edge list of reactions
}
\examples{
get_S_mats(all_mods, my_species, edge_list = T)
}
