% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{get_species_cmp_scores}
\alias{get_species_cmp_scores}
\title{Updated version of getting all single-species CMP scores for every compound and taxon}
\usage{
get_species_cmp_scores(species_table, network, normalize = T,
  relAbund = T, manual_agora = F, humann2 = F, leave_rxns = F,
  kos_only = F, remove_rev = T)
}
\arguments{
\item{species_table}{OTU abundance table (wide format)}

\item{network}{Species-specific network table, product of build_network functions}

\item{normalize}{Whether to normalize rows when making the network EMM}

\item{relAbund}{Whether to use relative abundance normalization}

\item{manual_agora}{Whether the metabolite are already specified using AGORA/BiGG (and therefore should not be mapped back to KEGG)}

\item{humann2}{Whether the species data is long-form humann2 gene-species abundances}

\item{leave_rxns}{Return individual abundance/direction scores for each species and rxn}

\item{kos_only}{Whether to call non-species-specific version of this function instead}

\item{remove_rev}{Whether to remove reversible reactions before calculating anything}
}
\value{
data.table of cmp scores for each taxon and compound
}
\description{
Updated version of getting all single-species CMP scores for every compound and taxon
}
\examples{
get_species_cmp_scores(species_data, network)
}
