% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{get_subset_picrust_ko_table}
\alias{get_subset_picrust_ko_table}
\title{Returns the melted picrust ko table corresponding to the given set of OTUs}
\usage{
get_subset_picrust_ko_table(
  otus,
  picrust_ko_table_directory,
  picrust_ko_table_suffix
)
}
\arguments{
\item{otus}{List of OTUs to find genomic content for}

\item{picrust_ko_table_directory}{Directory of PICRUSt genome OTU predictions}

\item{picrust_ko_table_suffix}{File naming of PICRUSt genome OTU predictions}
}
\value{
Genomic content for list of OTUs
}
\description{
Returns the melted picrust ko table corresponding to the given set of OTUs
}
\examples{
get_subset_picrust_ko_table(otus, picrust_dir, picrust_suffix)
}
