% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{get_text}
\alias{get_text}
\title{Get text constants used for MIMOSA interactive app from sysdata}
\usage{
get_text(text_name)
}
\arguments{
\item{text_name}{Variable name indicating desired text option}
}
\value{
named variable with its associated text
}
\description{
Get text constants used for MIMOSA interactive app from sysdata
}
\examples{
get_text("database_options")

}
