% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMOSA2.R
\name{getFDR}
\alias{getFDR}
\title{Get the q-values from a MIMOSA2 model}
\usage{
getFDR(x)
}
\arguments{
\item{x}{\code{MIMOSA2model} object}
}
\description{
Return the q-values from a MIMOSA2 model, calculated by the method of Newton et al.
}
