% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{getModelInfo}
\alias{getModelInfo}
\title{Returns the column from the picrust tables that corresponds to the genomic content of the indicated OTU}
\usage{
getModelInfo(matFile)
}
\arguments{
\item{out}{OTU/seq table}
}
\description{
Returns the column from the picrust tables that corresponds to the genomic content of the indicated OTU
}
\examples{
get_genomic_content_from_picrust_table(otu, picrust_dir, picrust_suffix)
}
