% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMOSA2.R
\name{getResponse}
\alias{getResponse}
\title{Get response calls from a MIMOSA2 model}
\usage{
getResponse(x, threshold = 0.01)
}
\arguments{
\item{x}{\code{MIMOSA2model} object}

\item{threshold}{\code{numeric} fdr threshold (default 0.01 i.e. 1%)
Return the response calls from a MIMOSA2 model at a given fdr threshold.}
}
\description{
Get response calls from a MIMOSA2 model
}
