% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{getWScores}
\alias{getWScores}
\title{Get Wilcoxon scores for Jaeckel's dispersion calculations}
\usage{
getWScores(nSamps)
}
\arguments{
\item{nSamps}{number of samples}
}
\value{
Vector of scores (length of nSamps)
}
\description{
Get Wilcoxon scores for Jaeckel's dispersion calculations
}
\examples{
getWScores(10)
}
