% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{gg_to_agora}
\alias{gg_to_agora}
\title{Returns AGORA species that a list of Greengenes OTUs were mapped to}
\usage{
gg_to_agora(otus,
  map_file_path = "data/rep_seqs/gg_13_8_99_toAGORA_97_map.txt")
}
\arguments{
\item{map_file_path}{File path to output of gg->agora vsearch mappings}

\item{species}{GG OTU abundance table}
}
\value{
List of AGORA species to use for model building
}
\description{
Returns AGORA species that a list of Greengenes OTUs were mapped to
}
\examples{
gg_to_agora(otu_list)
}
