% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{img_uri}
\alias{img_uri}
\title{Generate img URIs for putting plots in output table}
\usage{
img_uri(x)
}
\arguments{
\item{x}{File path to plot image}
}
\value{
Image URI
}
\description{
Generate img URIs for putting plots in output table
}
