% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{initialize}
\alias{initialize}
\title{Initialize model parameters}
\usage{
initialize(P, Ntot, ns1, nu1, ns0, nu0, K)
}
\arguments{
\item{P}{\code{numeric} The number of observations / subjects.}

\item{Ntot}{matrix of total counts}

\item{ns1}{vector of counts at time 1 stimulated}

\item{nu1}{vector of counts at time 1 unstimulated}

\item{ns0}{vector of counts at time 0 stimulated}

\item{nu0}{vector of counts at time 0 unstimulated}

\item{K}{\code{numeric} number of components in the model.}
}
\value{
\code{list} of initialized parameter estimates with components \code{inds} \code{theta_hat} \code{pi_est}.
}
\description{
Initialize model parameters
}
\details{
The \code{thetahat} are the initial hyperparameter estimates. The \code{pi_est} are the initial mixing proportions. The \code{inds} are the initial assignments of data to model components.
}
