% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{invlogit}
\alias{invlogit}
\title{invlogit}
\usage{
invlogit(x)
}
\arguments{
\item{x}{\code{numeric} between (-Inf,Inf)}
}
\value{
\code{invlogit(x)}
}
\description{
Return the inverse logit of x.
}
\details{
No bounds checking is done on the input
}
\examples{
invlogit(0)
}
