% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{j.disp.fit}
\alias{j.disp.fit}
\title{Jaeckel's dispersion from model predictions}
\usage{
j.disp.fit(yhat, y, scrs = NULL)
}
\arguments{
\item{yhat}{Model predictions}

\item{y}{True response values}

\item{scrs}{Optional score values (otherwise will use Wilcoxon)}
}
\value{
Value of Jaeckel's dispersion between predictions and the model
}
\description{
Jaeckel's dispersion from model predictions
}
\examples{
j.disp.fit(preds, y_vals)
}
