% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{j.disp.mean}
\alias{j.disp.mean}
\title{Jaeckel's dispersion from the mean (null)}
\usage{
j.disp.mean(y, scrs = NULL)
}
\arguments{
\item{y}{Vector of response values}

\item{scrs}{Optional score values (otherwise will use Wilcoxon)}
}
\value{
Value of Jaeckel's dispersion from the mean for this set of values
}
\description{
Jaeckel's dispersion from the mean (null)
}
\examples{
j.disp.mean(y_vals)
}
