% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{j.disp.rsq.adj}
\alias{j.disp.rsq.adj}
\title{Adjusted R-squared for a rank-based model}
\usage{
j.disp.rsq.adj(null_disp, model_disp, tauhat, df2, df1 = 1)
}
\arguments{
\item{null_disp}{Null dispersion value}

\item{model_disp}{Model dispersion value}

\item{tauhat}{Estimate of tau scale parameter}

\item{df2}{Degrees of freedom}

\item{df1}{Model df (assumes 1)}
}
\value{
Adjusted r-squared
}
\description{
Adjusted R-squared for a rank-based model
}
\examples{
j.disp.rsq.adj(null_disp, model_disp, tau, df2)
}
