% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{kegg_agora_mets}
\alias{kegg_agora_mets}
\title{Finds AGORA IDs for a list of KEGG metabolites (i.e. when adding to AGORA model)}
\usage{
kegg_agora_mets(kegg_ids)
}
\arguments{
\item{kegg_ids}{List of KEGG metabolite IDs}
}
\value{
vector of AGORA/Cobra IDs corresponding to agora_ids
}
\description{
Finds AGORA IDs for a list of KEGG metabolites (i.e. when adding to AGORA model)
}
\examples{
kegg_agora_mets(kegg_ids)
}
