% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{ll2}
\alias{ll2}
\title{Model component 2}
\usage{
ll2(par, Ntot, ns1, nu1, ns0, nu0)
}
\arguments{
\item{par}{vector of parameters.}

\item{Ntot}{\code{matrix} integer vector of total trials. One row per subject. Should have four columns named "ns1" "ns0" "nu1" and "nu0"}

\item{ns1}{\code{numeric} integer vector of successes in condition 1 treatment s.}

\item{nu1}{\code{numeric} integer vector of successes in condition 1 treatment u.}

\item{ns0}{\code{numeric} integer vector of successes in condition 0 treatment s.}

\item{nu0}{\code{numeric} integer vector of successes in condition 0 treatment u.}
}
\description{
Represents non-responders.
Log-Likelihood where all stims are equal to unstim.
}
\details{
The model hyperparameters are shared across treatments and conditions and are equal to the control.
}
\seealso{
\link{bbll} \link{MIMOSA2}
}

