% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{ll5}
\alias{ll5}
\title{Model component 5}
\usage{
ll5(par, Ntot, ns1, nu1, ns0, nu0)
}
\arguments{
\item{par}{vector of parameters.}

\item{Ntot}{\code{matrix} integer vector of total trials. One row per subject. Should have four columns named "ns1" "ns0" "nu1" and "nu0"}

\item{ns1}{\code{numeric} integer vector of successes in condition 1 treatment s.}

\item{nu1}{\code{numeric} integer vector of successes in condition 1 treatment u.}

\item{ns0}{\code{numeric} integer vector of successes in condition 0 treatment s.}

\item{nu0}{\code{numeric} integer vector of successes in condition 0 treatment u.}
}
\description{
Likelihood for response where stims at post-vaccine and no response at baseline mu[s0]
}
\details{
Represents a vaccine specific response where there is no response at condition 0, but a response at condition 1.
}
\seealso{
\link{bbll} \link{MIMOSA2}
}

