% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{load_agora_models}
\alias{load_agora_models}
\title{Reads a set of AGORA models from Matlab files}
\usage{
load_agora_models(mod_list, agora_path = "data/AGORA/")
}
\arguments{
\item{mod_list}{List of species/model IDs to load}

\item{agora_path}{Path to AGORA files}
}
\value{
List of model objects from files
}
\description{
Reads a set of AGORA models from Matlab files
}
\examples{
load_agora_models(species_list, "data/AGORA/")
}
