% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(p)
}
\arguments{
\item{p}{\code{numeric} between (0,1)}
}
\value{
logit(p)
}
\description{
Return the logit of p
}
\details{
No bounds checking is done on the input
}
\examples{
logit(0.5)
}

