% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{map_seqvar}
\alias{map_seqvar}
\title{Assign seq vars to OTUs or AGORA models using vsearch}
\usage{
map_seqvar(
  seqs,
  repSeqDir = "data/AGORA/",
  repSeqFile = "agora_NCBI_16S.udb",
  method = "vsearch",
  vsearch_path = "vsearch",
  file_prefix = "seqtemp",
  seqID = 0.99,
  add_agora_names = T,
  add_embl_names = F,
  otu_tab = F
)
}
\arguments{
\item{seqs}{vector of sequence variants}

\item{repSeqDir}{File path to reference database}

\item{repSeqFile}{File name with reference sequences}

\item{method}{Only "vsearch" currently implemented}

\item{vsearch_path}{Path to vsearch executable}

\item{file_prefix}{File name prefix for output}

\item{seqID}{threshold for vsearch --usearch-global search}

\item{add_agora_names}{Whether to add AGORA IDs to the table}

\item{add_embl_names}{Whether to add RefSeq/embl_gems names to the table}

\item{otu_tab}{Whether to return an OTU table, or just the matched sequences}
}
\value{
Table of alignment results (original sequence, hit ID)
}
\description{
Assign seq vars to OTUs or AGORA models using vsearch
}
\examples{
map_seqvar(seqs)
}
