% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{map_to_kegg}
\alias{map_to_kegg}
\title{Convert metabolite name table to KEGG metabolite table}
\usage{
map_to_kegg(met_table)
}
\arguments{
\item{met_table}{Table of metabolite abundances}
}
\value{
A new table of metabolite abundances using KEGG compound IDs
}
\description{
Convert metabolite name table to KEGG metabolite table
}
\examples{
new_mets = map_to_kegg(mets)
}
