% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{met_table_fix}
\alias{met_table_fix}
\title{Set column name to compound and filter metabolites with too many zeros or NAs}
\usage{
met_table_fix(met_table, nzero_filt = 5)
}
\arguments{
\item{met_table}{Metabolite table}

\item{nzero_filt}{Number of samples that must have a nonzero/non-NA value to include that metabolite}
}
\value{
Filtered metabolite table
}
\description{
Set column name to compound and filter metabolites with too many zeros or NAs
}
\examples{
met_table_fix(met_table)
}
