% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMOSA2.R
\name{MIMOSA2}
\alias{MIMOSA2}
\title{Fit a MIMOSA model with baseline.}
\usage{
MIMOSA2(Ntot,ns1,nu1,ns0,nu0)
}
\arguments{
\item{Ntot}{\code{matrix} integer vector of total trials. One row per subject. Should have four columns named "ns1" "ns0" "nu1" and "nu0"}

\item{ns1}{\code{numeric} integer vector of successes in condition 1 treatment s.}

\item{nu1}{\code{numeric} integer vector of successes in condition 1 treatment u.}

\item{ns0}{\code{numeric} integer vector of successes in condition 0 treatment s.}

\item{nu0}{\code{numeric} integer vector of successes in condition 0 treatment u.}

\item{tol}{\code{numeric} tolerance for stopping criteria, change in relative log-likelihood.}

\item{maxit}{\code{numeric} maximum number of iterations}
}
\value{
\code{list} of fitted model parameters with components \code{z} \code{inds} \code{thetahat} \code{pi_est}
}
\description{
Fit a MIMOSA model with baseline.
}
\seealso{
\link{ORTest} \link{ROC} \link{ROCPlot} \link{Boxplot} \link{simulate_MIMOSA2} \link{logit} \link{invlogit}
}

