% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TestOR.R
\name{ORTest}
\alias{ORTest}
\title{Test for a difference of odds ratios}
\usage{
ORTest(Ntot, ns1, nu1, ns0, nu0)
}
\arguments{
\item{Ntot}{\code{matrix} integer vector of total trials. One row per subject. Should have four columns named "ns1" "ns0" "nu1" and "nu0"}

\item{ns1}{\code{numeric} integer vector of successes in condition 1 treatment s.}

\item{nu1}{\code{numeric} integer vector of successes in condition 1 treatment u.}

\item{ns0}{\code{numeric} integer vector of successes in condition 0 treatment s.}

\item{nu0}{\code{numeric} integer vector of successes in condition 0 treatment u.}
}
\value{
\code{numeric} vector of one-sided p-values from a standard normal.
}
\description{
Test for a difference of odds ratios
}
\details{
SE of log odds = sqrt(1/n1+1/n2+1/n3+1/n4)
SE of difference in log odds  = sqrt(SE1^2 +SE2^2)
delta  = difference in log odds
Test delta / SE from standard normal.
}
\examples{
s = simulate_MIMOSA2()
R = MIMOSA2(Ntot=s$Ntot, ns1 = s$ns1, nu1 = s$nu1, nu0 = s$nu0, ns0 = s$ns0)
or_result = ORTest(Ntot = s$Ntot, ns1 = s$ns1, nu1 = s$nu1, nu0 = s$nu0, ns0 = s$ns0)
}
\seealso{
\link{MIMOSA2}
}

