% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{otus_to_agora}
\alias{otus_to_agora}
\title{Returns AGORA species that a list of Greengenes OTUs were mapped to}
\usage{
otus_to_agora(otus, database = "Greengenes",
  gg_file_path = "data/rep_seqs/gg_13_8_99_toAGORA_97_map.txt",
  silva_file_path = "data/rep_seqs/silva_132_99_toAGORA_97_map.txt")
}
\arguments{
\item{otus}{GG or SILVA OTU abundance table}

\item{database}{Greengenes or SILVA}

\item{gg_file_path}{File path to output of gg->agora vsearch mappings}

\item{silva_file_path}{File path to output of silva->agora vsearch mappings}
}
\value{
List of AGORA species to use for model building
}
\description{
Returns AGORA species that a list of Greengenes OTUs were mapped to
}
\examples{
otus_to_agora(otu_list)
}
