% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{otus_to_db}
\alias{otus_to_db}
\title{Returns AGORA species that a list of Greengenes OTUs were mapped to}
\usage{
otus_to_db(
  otus,
  database = "Greengenes",
  target_db = "AGORA",
  data_prefix = "data/OTU_model_mapping_files/",
  gg_file_prefix = "gg_13_8_99",
  silva_file_prefix = "silva_132_99",
  agora_file_suffix = "_toAGORA_97_map.txt",
  refseq_file_suffix = "_toRefSeq_97_map.txt"
)
}
\arguments{
\item{otus}{GG or SILVA OTU abundance table}

\item{database}{Greengenes or SILVA}

\item{target_db}{AGORA or RefSeq}

\item{gg_file_prefix}{File path to output of gg->db vsearch mappings}

\item{silva_file_prefix}{File path to output of silva->db vsearch mappings}

\item{agora_file_suffix}{File path to output of xx-> agora vsearch mappings}

\item{refseq_file_suffix}{File path to output of xx->refseq vsearch mappings}
}
\value{
List of AGORA species to use for model building
}
\description{
Returns AGORA species that a list of Greengenes OTUs were mapped to
}
\examples{
otus_to_db(otu_list)
}
