% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{plot_all_cmp_mets}
\alias{plot_all_cmp_mets}
\title{Generate a list of plots of CMP values vs metabolite concentrations for all metabolites}
\usage{
plot_all_cmp_mets(cmp_table, met_table, mod_results)
}
\arguments{
\item{cmp_table}{Dataset of CMP scores (long format, can be species-specific or added already) for a single metabolite}

\item{met_table}{Dataset of metabolite concentrations (long format) for a single metabolite}

\item{mod_results}{Optional additional info on model comparison results}

\item{sample_col}{Column name for sample column (shared between the tables)}
}
\value{
plot of contributions
}
\description{
Generate a list of plots of CMP values vs metabolite concentrations for all metabolites
}
\examples{
plot_all_cmp_mets(varShares)
}
