% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{plot_contributions}
\alias{plot_contributions}
\title{Plot species contributions for each metabolite}
\usage{
plot_contributions(varShares, numrow = 10)
}
\arguments{
\item{varShares}{Dataset of contributions}
}
\value{
plot of contributions
}
\description{
Plot species contributions for each metabolite
}
