% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{plot_summary_contributions}
\alias{plot_summary_contributions}
\title{Plot summary of metabolite-species contributions}
\usage{
plot_summary_contributions(varShares, include_zeros = T)
}
\arguments{
\item{varShares}{Dataset of contributions}

\item{include_zeros}{Whether to plot taxa that do not have any contribution}
}
\value{
plot of contributions
}
\description{
Plot summary of metabolite-species contributions
}
\examples{
plot_summary_contributions(varShares)
}
