% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{randomString}
\alias{randomString}
\title{Get random string (for saving temp files)}
\usage{
randomString()
}
\value{
5 random letters and 5 random numbers
}
\description{
Get random string (for saving temp files)
}
\examples{
randomString()

}
