% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{rank_based_rsq_contribs}
\alias{rank_based_rsq_contribs}
\title{Calculates contributions to variance for Rfit regression models}
\usage{
rank_based_rsq_contribs(species_cmps, mets_melt, config_table,
  cmp_mods = NULL, adj_rsq = F, nperm = 500,
  signif_threshold = 0.01, return_perm = F, merge_low_abund = NULL)
}
\arguments{
\item{species_cmps}{Table of species contributions (without residuals)}

\item{mets_melt}{Table of metabolite concentrations}

\item{config_table}{Configuration table including model-fitting settings}

\item{cmp_mods}{Output of fit_cmp_mods}

\item{adj_rsq}{Whether to decompose standard disp rsq or adjusted}

\item{nperm}{Number of permutations per decomposition}

\item{signif_threshold}{Will only analyze models with a p-value below this threshold}

\item{return_perm}{Whether to return full table of permutation results}

\item{merge_low_abund}{Optional list of low-abundance species, will be merged into single "Other" taxon}
}
\value{
Data table of variance shares by each species in the original table for each metabolite
}
\description{
Calculates contributions to variance for Rfit regression models
}
\examples{
rank_based_rsq_contribs(species_cmps, mets_melt, config_table)
}
