% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{read_mimosa2_files}
\alias{read_mimosa2_files}
\title{Read in files for a MIMOSA 2 analysis}
\usage{
read_mimosa2_files(config_table)
}
\arguments{
\item{config_table}{Data.table of configuration settings, including paths to abundance data files}
}
\value{
list of 2 processed abundance data.tables
}
\description{
Read in files for a MIMOSA 2 analysis
}
\examples{
read_mimosa2_files(config_table)
}
