% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TestOR.R
\name{ROC}
\alias{ROC}
\title{Generate matrix for ROC plot.}
\usage{
ROC(or_test, z, truth)
}
\arguments{
\item{or_test}{output of \code{ORTest}}

\item{z}{output of MIMOSA2 \code{z} component}

\item{truth}{\code{vector} of ground truth of length number of observations where response = "R".}
}
\value{
\code{matrix} with columns "FPR" "TPR" and "Method"
}
\description{
Generate matrix for ROC plot.
}
\examples{
s = simulate_MIMOSA2()
R = MIMOSA2(Ntot=s$Ntot, ns1 = s$ns1, nu1 = s$nu1, nu0 = s$nu0, ns0 = s$ns0)
or_result = ORTest(Ntot = s$Ntot, ns1 = s$ns1, nu1 = s$nu1, nu0 = s$nu0, ns0 = s$ns0)
ROC(or_test = or_result, z = R$z, truth = s$truth)
}
\seealso{
\link{MIMOSA2}
}

