% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_all_metabolites_FBA2}
\alias{run_all_metabolites_FBA2}
\title{Run full MIMOSA1 analysis on simulation data (2019)}
\usage{
run_all_metabolites_FBA2(
  run_prefix,
  fake_spec,
  fake_mets,
  network,
  spec_codes,
  cor_method = "spearman",
  correction = "fdr",
  nperm = 10000,
  nonzero_filter = 4,
  kegg_translate = ""
)
}
\arguments{
\item{run_prefix}{}

\item{fake_spec}{processed species table}

\item{fake_mets}{processed metabolite table}

\item{network}{output of build_metabolic_model}

\item{spec_codes}{Table of species IDs}

\item{cor_method}{spearman or pearson}

\item{correction}{fdr or bonferroni}

\item{nperm}{Number of Mantel permutations}

\item{nonzero_filter}{Number of samples that must be nonzero for a metabolite to be included}

\item{kegg_translate}{Compound ID table}
}
\value{
Output of run_all_metabolites
}
\description{
Run full MIMOSA1 analysis on simulation data (2019)
}
\examples{
run_all_metabolites_FBA2("sim1", species, mets, network, spec_codes)
}
