% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_mimosa1}
\alias{run_mimosa1}
\title{Run full MIMOSA1 analysis (2019)}
\usage{
run_mimosa1(
  species,
  mets,
  species_file = "mimosa1",
  picrust_file_paths = c("data/picrustGenomeData/16S_13_5_precalculated.tab.gz",
    "data/picrustGenomeData/indivGenomes/", "_genomic_content.tab"),
  rxn_path = "data/KEGG/full_rxn_table.txt",
  simulated = F,
  config_table = NULL
)
}
\arguments{
\item{species}{processed species table}

\item{mets}{processed metabolite table}

\item{species_file}{File prefix from species data or other ID for saving}

\item{picrust_file_paths}{paths to picrust1 files}

\item{rxn_path}{Path to processed mapformula reaction table}

\item{simulated}{Whether this is simulated data}

\item{config_table}{A MIMOSA2 configuration table of analysis settings (required only for using simulation data)}
}
\value{
Output of run_all_metabolites
}
\description{
Run full MIMOSA1 analysis (2019)
}
\examples{
run_mimosa1(species, mets)
}
