% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{run_mimosa2}
\alias{run_mimosa2}
\title{Run a MIMOSA 2 analysis}
\usage{
run_mimosa2(config_table)
}
\arguments{
\item{config_table}{Data.table of input files and settings for MIMOSA analysis}
}
\value{
Scaling model and variance contribution results
}
\description{
Run a MIMOSA 2 analysis
}
\examples{
run_mimosa2(species_file, met_file, config_file)
}
