% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{run_samp_shapley_analysis}
\alias{run_samp_shapley_analysis}
\title{Calculates contributions to variance using permutation/subset Shapley analysis for regression fits}
\usage{
run_samp_shapley_analysis(
  species_cmps,
  mets_melt,
  config_table,
  nperm = 15000,
  signif_threshold = 0.01,
  return_perm = F
)
}
\arguments{
\item{species_cmps}{Table of species contributions (without residuals)}

\item{mets_melt}{Table of metabolite concentrations}

\item{config_table}{Configuration table including model-fitting settings}

\item{nperm}{Number of permutations per decomposition}

\item{signif_threshold}{Will only analyze models with a p-value below this threshold}

\item{return_perm}{Whether to return full table of permutation results}
}
\value{
Data table of variance shares by each species in the original table for each metabolite
}
\description{
Calculates contributions to variance using permutation/subset Shapley analysis for regression fits
}
\examples{
run_samp_shapley_analysis(species_cmps)
}
