% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_MIMOSA2}
\alias{simulate_MIMOSA2}
\title{Simulate data for MIMOSA2 model, ICS with baseline.}
\usage{
simulate_MIMOSA2(effect = 5e-04, phi = 10000, P = 100)
}
\arguments{
\item{effect}{\code{numeric} effect size for ps1 - pu1 - ps0 - pu0}

\item{phi}{\code{numeric} integer, the precision.}

\item{P}{\code{numeric} number of subjects.}

\item{seed}{\code{numeric} random seed}
}
\value{
\code{list} with components "Ntot" "ns0" "nu0" "ns1" "nu1" "truth"
}
\description{
Simulate data for MIMOSA2 model, ICS with baseline.
}
\examples{
s = simulate_MIMOSA2()
}
\seealso{
\link{MIMOSA2}
}

