% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{spec_table_fix}
\alias{spec_table_fix}
\title{Set column name of OTU table to "OTU"}
\usage{
spec_table_fix(species_table)
}
\value{
Species table with column name as OTU
}
\description{
Set column name of OTU table to "OTU"
}
\examples{
spec_table_fix(spec_table)
}
