% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{test_m2_analysis}
\alias{test_m2_analysis}
\title{Run toy analysis to test package installation}
\usage{
test_m2_analysis(test_vsearch = F)
}
\arguments{
\item{test_vsearch}{Whether to test installation of vsearch}
}
\value{
None
}
\description{
Run toy analysis to test package installation
}
\examples{
test_m2_analysis()
}
