% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{transform_cmps}
\alias{transform_cmps}
\title{Apply a transformation to CMP data}
\usage{
transform_cmps(cmp_dat, score_transform)
}
\arguments{
\item{cmp_dat}{Dataset of species-specific metabolic potential scores}

\item{score_transform}{transformation to apply (current options: zscore, log1plus, sqrt)}
}
\value{
Dataset with transformed values (rawValue is pre-transofrmation, value is now transformed value)
}
\description{
Apply a transformation to CMP data
}
\examples{
transform_cmps(indiv_cmps, met_transform = "zscore")
}
