% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{transform_mets}
\alias{transform_mets}
\title{Apply a transformation to metabolite data}
\usage{
transform_mets(met_dat, met_transform)
}
\arguments{
\item{met_dat}{Dataset of metabolite concentrations}

\item{met_transform}{transformation to apply (current options: zscore, log1plus, sqrt)}
}
\value{
Dataset with transformed values (rawValue is pre-transofrmation, value is now transformed value)
}
\description{
Apply a transformation to metabolite data
}
\examples{
transform_mets(mets_melt, met_transform = "zscore")
}
