% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{var_shares_cmps}
\alias{var_shares_cmps}
\title{Get variance share contributions to CMP scores for all compounds}
\usage{
var_shares_cmps(cmps_sub_good, all_rxns, subjects, norm_kos, ko_net,
  all_taxa, single_spec_cmps, contrib_threshold = 0.45)
}
\arguments{
\item{cmps_sub_good}{CMP scores for metabolites with abundance data}

\item{all_rxns}{list of relevant reactions for each metabolite}

\item{subjects}{vector of subjects}

\item{norm_kos}{data.table of gene abundances}

\item{ko_net}{network created by generate_genomic_network}

\item{all_taxa}{vector of OTUs}

\item{single_spec_cmps}{single-species CMP scores calculated from get_spec_contribs function}

\item{contrib_threshold}{Threshold for binary contribution}
}
\value{
data.table of variance contributions to CMP scores
}
\description{
Get variance share contributions to CMP scores for all compounds
}
\examples{
var_shares_cmps(cmps_total, all_rxns, subjects, norm_kos, ko_net, all_taxa, single_spec_cmps, contrib_threshold)
}
