% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{build_species_networks_w_agora}
\alias{build_species_networks_w_agora}
\title{Taxa mapping functions, MIMOSA2
June 2018
Finds close AGORA models and then imports those for each species. Builds PICRUSt-based network for remaining species}
\usage{
build_species_networks_w_agora(mod_list, usePreprocessed = T,
  agora_path = "data/AGORA/", remove_sinks = T, add_rev = T)
}
\arguments{
\item{usePreprocessed}{Whether to process .mat files or use pre-computed versions}

\item{agora_path}{Path to AGORA models}

\item{remove_sinks}{Remove reactions that have NA as a substrate or product (usually uninformative transporters etc)}

\item{add_rev}{Whether to add reverse of reversible reactions (default = T)}

\item{species_list}{List of species from mapping}
}
\value{
A list with 2 entries - one a species abundance table in terms of the new species, the second a table of reactions for each species
}
\description{
Taxa mapping functions, MIMOSA2
June 2018
Finds close AGORA models and then imports those for each species. Builds PICRUSt-based network for remaining species
}
\examples{
build_species_networks_w_agora(species_table, "Greengenes 13_5 or 13_8", simThreshold = 0.99)
}
