% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{filter_currency_metabolites}
\alias{filter_currency_metabolites}
\title{Remove metabolites linked to many reactions from a metabolic network.}
\usage{
filter_currency_metabolites(rxn_table, degree_filter = 30)
}
\arguments{
\item{rxn_table}{Network table of reactions}

\item{degree_filter}{Compounds involved in more than this number of reactions will be removed}
}
\value{
Network table without compounds found in reactions greater than the cutoff
}
\description{
Remove metabolites linked to many reactions from a metabolic network.
}
\examples{
filter_currency_metabolites(rxn_table, degree_filter = 50)
}
