% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{fit_cmp_mods}
\alias{fit_cmp_mods}
\title{Fits model scaling total CMPs to metabolite concentrations}
\usage{
fit_cmp_mods(species_cmps, fake_mets_melt)
}
\arguments{
\item{species_cmps}{Table of species contribution abundances}

\item{fake_mets_melt}{Table of metabolite concentrations}
}
\value{
List of 2 data.tables - one with model summary results, one with model residuals
}
\description{
Fits model scaling total CMPs to metabolite concentrations
}
\examples{
fit_cmp_mods(species_cmps, met_data)
}
