% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{fit_cmp_net_edit}
\alias{fit_cmp_net_edit}
\title{Greedy algorithm to test network improvements /refine rxn network}
\usage{
fit_cmp_net_edit(
  network,
  species,
  mets_melt,
  manual_agora = F,
  rsq_factor = 1.15,
  min_rsq = 0.1,
  min_rxns = 3,
  max_rxns_test = 40,
  rank_based = F,
  species_specific = T
)
}
\arguments{
\item{network}{Edge list reaction table (including reversibility info)}

\item{species}{Table of species abundances}

\item{mets_melt}{Table of metabolite concentrations}

\item{manual_agora}{Whether to leave in agora/bigg format}

\item{rsq_factor}{Improvement in R squared needed to keep going}

\item{min_rsq}{Minimum r squared needed to accept an improvement}

\item{min_rxns}{Minimum number of reactions remaining to continue iterating}

\item{rank_based}{Whether to use rank-based regression}

\item{species_specific}{Whether to adjust rxns in a species-specific or whole-community manner}

\item{max_rxns}{Maximum number of reactions to try}
}
\value{
A list of model fitting results as well as revised versions of the community metabolic network model and the updated CMP scores
}
\description{
Greedy algorithm to test network improvements /refine rxn network
}
\examples{
get_best_rxn_subset(met1, species, met_data)
}
