% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{get_otus_from_seqvar}
\alias{get_otus_from_seqvar}
\title{Assign seq vars to OTUs or AGORA models using vsearch}
\usage{
get_otus_from_seqvar(seqs,
  repSeqDir = "~/Documents/MIMOSA2shiny/data/blastDB/",
  repSeqFile = "agora_NCBI_16S.fna", method = "vsearch",
  file_prefix = "seqtemp", seqID = 0.99, add_agora_names = T)
}
\arguments{
\item{seqs}{vector of sequence variants}

\item{repSeqDir}{File path to reference database}
}
\value{
Table of alignment results (original sequence, hit ID)
}
\description{
Assign seq vars to OTUs or AGORA models using vsearch
}
\examples{
get_otus_from_seqvar(seqs)
}
