% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{get_species_cmp_scores}
\alias{get_species_cmp_scores}
\title{Updated version of getting all single-species CMP scores for every compound and taxon}
\usage{
get_species_cmp_scores(species_table, network, normalize = T,
  relAbund = T)
}
\arguments{
\item{species_table}{OTU abundance table (wide format)}

\item{network}{Species-specific network table, product of build_network functions}

\item{normalize}{Whether to normalize rows when making the network EMM}
}
\value{
data.table of cmp scores for each taxon and compound
}
\description{
Updated version of getting all single-species CMP scores for every compound and taxon
}
\examples{
get_species_cmp_scores(species_data, network)
}
