% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_agora.R
\name{getModelInfo}
\alias{getModelInfo}
\title{Reads metabolic model in Cobra/Matlab format}
\usage{
getModelInfo(matFile)
}
\arguments{
\item{matFile}{Path to matlab-formatted AGORA model}
}
\value{
Returns matlab object contents from the readMat function
}
\description{
Reads metabolic model in Cobra/Matlab format
}
\examples{
getModelInfo(otu_file)
}
