% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Initialization.R
\name{initialize}
\alias{initialize}
\title{Initialize model parameters}
\usage{
initialize(P, Ntot, ns1, nu1, ns0, nu0, random = FALSE, K)
}
\arguments{
\item{P}{\code{numeric} The number of observations / subjects.}

\item{ns1}{}

\item{nu1}{}

\item{ns0}{}

\item{nu0}{}

\item{Nott}{}
}
\value{
\code{list} of initialized parameter estimates with components \code{inds} \code{theta_hat} \code{pi_est}.
}
\description{
Initialize model parameters
}
\details{
The \code{thetahat} are the initial hyperparameter estimates. The \code{pi_est} are the initial mixing proportions. The \code{inds} are the initial assignments of data to model components.
}

