% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{read_mimosa2_files}
\alias{read_mimosa2_files}
\title{Read in files for a MIMOSA 2 analysis}
\usage{
read_mimosa2_files(file_list, configTable, app = T)
}
\arguments{
\item{file_list}{List of shiny file inputs to load}

\item{configTable}{Table of configuration parameters}

\item{app}{Whether configurations are coming from the Shiny app or otherwise}
}
\value{
list of processed abundance data.tables
}
\description{
Read in files for a MIMOSA 2 analysis
}
\examples{
read_mimosa2_files(input_file_list, config_table)
}
