% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TestOR.R
\name{ROCPlot}
\alias{ROCPlot}
\title{Plot ROCs using ggplot}
\usage{
ROCPlot(R, lambda = 1)
}
\arguments{
\item{R}{\code{matrix} output of \code{ROC()}}

\item{lambda}{\code{numeric} smoothing. Default 1}
}
\description{
Plot ROCs using ggplot
}
\details{
uses the output of ROC
}
\examples{
s = simulate_MIMOSA2()
R = MIMOSA2(Ntot=s$Ntot, ns1 = s$ns1, nu1 = s$nu1, nu0 = s$nu0, ns0 = s$ns0)
or_result = ORTest(Ntot = s$Ntot, ns1 = s$ns1, nu1 = s$nu1, nu0 = s$nu0, ns0 = s$ns0)
roc = ROC(or_test = or_result, z = R$z, truth = s$truth)
ROCPlot(roc)
}
\seealso{
\link{MIMOSA2}
}

