% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_mimosa2_funcs.R
\name{run_mimosa2}
\alias{run_mimosa2}
\title{Run a MIMOSA 2 analysis}
\usage{
run_mimosa2(config_table, species = "", mets = "", make_plots = F,
  save_plots = F, plot_dir = "mimosa2plots")
}
\arguments{
\item{config_table}{Data.table of input files and settings for MIMOSA analysis OR path to such a table}

\item{species}{Optionally provide already-read-in species data}

\item{mets}{Optionally provide already-read-in metabolite data}

\item{make_plots}{Whether to generate plots for each metabolite and extended output (as provided via the web server)}

\item{save_plots}{Whether to save plots to files}

\item{plot_dir}{Path for saving plots}
}
\value{
Scaling model and variance contribution results
}
\description{
Run a MIMOSA 2 analysis
}
\examples{
run_mimosa2(config_table, species, mets)
}
