% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_MIMOSA2}
\alias{simulate_MIMOSA2}
\title{Simulate data for MIMOSA2 model, ICS with baseline.}
\usage{
simulate_MIMOSA2(effect = 5e-04, bg_effect = 0,
  baseline_stim_effect = 0.00025, baseline_background = 1e-04, phi = 5000,
  P = 100)
}
\arguments{
\item{effect}{\code{numeric} effect size for ps1 - pu1 - ps0 + pu0. Default 5e-4}

\item{bg_effect}{\code{numeric} effect size for background pu0-pu1. Default 0}

\item{baseline_stim_effect}{\code{numeric} baseline stimulation effect. Default 2.5e-4}

\item{baseline_background}{\code{numeric} baseline background effect. Default 1e-4}

\item{phi}{\code{numeric} integer, the precision. Default 5000}

\item{P}{\code{numeric} number of subjects.'}
}
\value{
\code{list} with components "Ntot" "ns0" "nu0" "ns1" "nu1" "truth"
}
\description{
Simulate data for MIMOSA2 model, ICS with baseline.
}
\examples{
s = simulate_MIMOSA2()
Run 100 simulations, dropping any errors (i.e. simulations that error out)
proportion of vaccine specific responses
Total observations simulated
from uniform with between 100,000 and 150,000 cells
Hyperprior mean for stimulated time 0
Non Stimulated post vaccine
Mean of hyperprior (stimulated time 1) $\\alpha/(\\alpha+\\beta)$
Precision of hyperprior
$\\alpha+\\beta = \\phi$
There are 8 model components.
1. all different
2. s0=u0
3. s1 = s0
4. u1 = u0
5. s0 = u0, s1 = u1
6. s1 = u1
7. s1 = u1 = s0 = u0
8. s1 = s0, u1 = u0
number of responders / non-responders / non-specific responders
Simulate from component 1
Simulate from component 2
Simulate from component 3.
Simulate from component 4.
simulate from component 5
component 6
component 7
component 8
Simulate count observations from binomial.
Empirical estimates of proportions
True response categories
}
\seealso{
\link{MIMOSA2}
}

