% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_MIMOSA2}
\alias{simulate_MIMOSA2}
\title{Simulate data for MIMOSA2 model, ICS with baseline.}
\usage{
simulate_MIMOSA2(effect = 5e-04, bg_effect = 0,
  baseline_stim_effect = 0.00025, baseline_background = 1e-04,
  phi = 5000, P = 100, rng = c(1e+05, 150000))
}
\arguments{
\item{effect}{\code{numeric} effect size for ps1 - pu1 - ps0 + pu0. Default 5e-4}

\item{bg_effect}{\code{numeric} effect size for background pu0-pu1. Default 0}

\item{baseline_stim_effect}{\code{numeric} baseline stimulation effect. Default 2.5e-4}

\item{baseline_background}{\code{numeric} baseline background effect. Default 1e-4}

\item{phi}{\code{numeric} integer, the precision. Default 5000}

\item{P}{\code{numeric} number of subjects.'}

\item{rng}{\code{numeric} vector of length 2. The range of total cell counts to simulate from.}
}
\value{
\code{list} with components "Ntot" "ns0" "nu0" "ns1" "nu1" "truth"
}
\description{
Simulate data for MIMOSA2 model, ICS with baseline.
}
\examples{
s = simulate_MIMOSA2()
}
\seealso{
\link{MIMOSA2}
}
