% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesPicrustFunctions.R
\name{spec_table_fix}
\alias{spec_table_fix}
\title{Set column name of OTU table to "OTU", and remove taxa with no abundance}
\usage{
spec_table_fix(species_table)
}
\arguments{
\item{species_table}{Species/OTU table}
}
\value{
Species table with column name as OTU
}
\description{
Set column name of OTU table to "OTU", and remove taxa with no abundance
}
\examples{
spec_table_fix(spec_table)
}
