% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{sumcll}
\alias{sumcll}
\title{Calculate the complete data log-likelihood across all observations.}
\usage{
sumcll(..., z, pi_est)
}
\arguments{
\item{...}{all model parameters and data (par, Ntot, ns0,ns1,nu0,nu1)}

\item{z}{\code{matrix} of type \code{numeric} represents the posterior probabilities of the class assignments of each observation to each component. Rows sum to 1.}

\item{pi_est}{\code{numeric} the mixing proportions.}
}
\description{
Calculate the complete data log-likelihood across all observations.
}
\seealso{
\link{bbll} \link{MIMOSA2}
}

