% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{sumcll}
\alias{sumcll}
\title{Sum of the complete data log-likelihood across all observations.}
\usage{
sumcll(..., inds, pi_est)
}
\arguments{
\item{...}{all model parameters and data (par, Ntot, ns0,ns1,nu0,nu1)}

\item{inds}{\code{matrix} of type \code{numeric} represents the max(z's), i.e. the class assignments of each observation to each component.}

\item{pi_est}{\code{numeric} the mixing proportions.}
}
\description{
Sum of the complete data log-likelihood across all observations.
}
\seealso{
\link{bbll} \link{MIMOSA2}
}

